+++ lib/src/codecs/codecs.dart
@@ -9,3 +9,4 @@ export 'ncname/ncname_codec.dart';
 export 'short/short_codec.dart';
 export 'byte/byte_codec.dart';
 export 'base64_binary/base64_binary_codec.dart';
+export 'hex_binary/hex_binary_codec.dart';