+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - Implement `xsd:short` data type codec
 - Implement `xsd:byte` data type codec
 - Implement `xsd:base64Binary` data type codec
+- Implement `xsd:hexBinary` data type codec